+++ library/Oidc/UserTable.php
@@ -10,7 +10,9 @@ use DateTimeZone;
 use Icinga\Authentication\Auth;
 use Icinga\Module\Oidc\Model\User;
 use Icinga\Web\Url;
+use ipl\Html\Html;
 use ipl\Orm\Model;
+use ipl\Web\Widget\Icon;
 
 /**
  * Table widget to display a list of Users
@@ -18,8 +20,8 @@ use ipl\Orm\Model;
 class UserTable extends DataTable
 {
     protected $defaultAttributes = [
-        'class'            => 'usage-table common-table table-row-selectable',
-        'data-base-target' => '_next'
+        'class'            => 'usage-table common-table',
+
     ];
 
     public function createColumns()
@@ -44,6 +46,15 @@ class UserTable extends DataTable
                     continue;
                 }
 
+                if ($column === "name"){
+                    $columns[$column."_link"] = [
+                        'label'  => $options['label']??$column,
+                        'column' => function ($data) use ($column) {
+                             return Html::tag('a',['data-base-target'=>'_next',"href"=>\ipl\Web\Url::fromPath("user/show", ['backend'=>$data->provider->name, 'user'=>$data->name])],$data->name);
+                        }
+                    ];
+                    continue;
+                }
 
                 if ($fieldtype === "autocomplete" || $fieldtype === "text" || $fieldtype === "select") {
                     $columns[$column] = $options['label']??$column;
@@ -70,6 +81,26 @@ class UserTable extends DataTable
                 $columns[$column] = $options;
             }
         }
+        if (Auth::getInstance()->hasPermission('oidc/user/modify')) {
+            $columns['action'] = [
+                'label' => mt('packagemirror', 'Action'),
+                'attributes' => ['class' => 'icon-col'],
+                'column' => function ($data) {
+                    return $data;
+                },
+                'renderer' => function ($data) {
+                    $div=Html::tag("div",['class'=>'action-column']);
+
+                    $icon=  new Icon('pencil', ['title' => mt('oidc', 'Edit')]);
+                    $a = Html::tag("a",['data-icinga-modal' => true, 'data-no-icinga-ajax' => true, 'class'=>'action-item','href'=> \ipl\Web\Url::fromPath('oidc/user/edit',['id'=>$data->id])]);
+                    $a->add($icon);
+                    $div->add($a);#
+
+
+                    return $div;
+                }
+            ];
+        }
 
 
         return $columns;
@@ -79,14 +110,6 @@ class UserTable extends DataTable
     protected function renderRow(Model $row)
     {
         $tr = parent::renderRow($row);
-
-        if (Auth::getInstance()->hasPermission('oidc/user/modify')) {
-            $url = Url::fromPath('oidc/user/edit', ['id' => $row->id]);
-
-            $tr->getFirst("td")->getAttributes()->add(['href' => $url->getAbsoluteUrl(), 'data-icinga-modal' => true,
-                'data-no-icinga-ajax' => true]);
-
-        }
 
         return $tr;
     }