+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## What's New
 
+### What's New in Version 0.6.9
+
+* update docs
+* add option to enforce https scheme behind reverse proxy for redirect urls
+* tested with keycloak and php 8.3.6
+
+### What's New in Version 0.6.8
+
+* moved edit function to action column (edit button)
+* usernames and groupnames now link to icingaweb2 access control.
+
 ### What's New in Version 0.6.7
 
 * experimental postgres support